Moved build expression unmarshaling tests into build script tests.
More refactoring and shuffling things around.
