+++ pkg/buildscript/testdata/buildexpression-alternate.json
