+++ pkg/buildscript/testdata/buildexpression-complex.json
