+++ pkg/buildscript/testdata/buildexpression-installer-complex.json
