+++ pkg/buildscript/testdata/buildexpression-installer.json
