+++ pkg/buildscript/testdata/buildexpression-nested.json
