+++ pkg/buildscript/testdata/buildexpression-new-objects.json
